refactor : UploadView 생성하여 AddView에 상속
